FLASKPROJEM/requirements.txt
Flask>=2.0
Pillow>=9.0
pytest>=7.0
